refactor(components): detach logo
- moved logo as separate component
